requirements.txt
flask
flask-cors
gunicorn
requests
psycopg2-binary